Don't redownload /unpack cached files

+++ moss/src/package/cache.rs
@@ -4,9 +4,9 @@
 
 use std::{io, path::PathBuf};
 
-use futures::StreamExt;
-use rayon::prelude::{IntoParallelRefIterator, ParallelIterator};
-use stone::read::Payload;
+use futures::{stream, StreamExt};
+use rayon::prelude::{IntoParallelIterator, ParallelIterator};
+use stone::{payload, read::Payload};
 use thiserror::Error;
 use tokio::{
     fs::{self, File},
@@ -20,6 +20,8 @@ use crate::{
     package::{Id, Meta},
     request, Installation,
 };
+
+const CONCURRENT_FS_TASKS: usize = 16;
 
 #[derive(Debug, Clone, Copy)]
 pub struct Progress {
@@ -43,9 +45,18 @@ pub async fn fetch(
     let url = meta.uri.as_ref().ok_or(Error::MissingUri)?.parse::<Url>()?;
     let hash = meta.hash.as_ref().ok_or(Error::MissingHash)?;
 
-    let mut bytes = request::get(url).await?;
-
     let download_path = download_path(installation, hash).await?;
+
+    if fs::try_exists(&download_path).await? {
+        return Ok(Download {
+            id: meta.id().into(),
+            path: download_path,
+            installation: installation.clone(),
+            was_cached: true,
+        });
+    }
+
+    let mut bytes = request::get(url).await?;
     let mut out = File::create(&download_path).await?;
 
     let mut total = 0;
@@ -69,6 +80,7 @@ pub async fn fetch(
         id: meta.id().into(),
         path: download_path,
         installation: installation.clone(),
+        was_cached: false,
     })
 }
 
@@ -77,6 +89,7 @@ pub struct Download {
     id: Id,
     path: PathBuf,
     installation: Installation,
+    pub was_cached: bool,
 }
 
 /// Upon fetch completion we have this unpacked asset bound with
@@ -145,6 +158,20 @@ impl Download {
             let mut reader = stone::read(File::open(&self.path)?)?;
 
             let payloads = reader.payloads()?.collect::<Result<Vec<_>, _>>()?;
+            let indicies = payloads
+                .iter()
+                .filter_map(Payload::index)
+                .flatten()
+                .collect::<Vec<_>>();
+
+            // If download was cached & all assets exist, we can skip unpacking
+            if self.was_cached && rt.block_on(check_assets_exist(&indicies, &self.installation)) {
+                return Ok(UnpackedAsset {
+                    id: self.id.clone(),
+                    payloads,
+                });
+            }
+
             let content = payloads
                 .iter()
                 .find_map(Payload::content)
@@ -161,10 +188,8 @@ impl Download {
                 &mut ProgressWriter::new(&content_file, content.plain_size, &on_progress),
             )?;
 
-            payloads
-                .par_iter()
-                .filter_map(Payload::index)
-                .flatten()
+            indicies
+                .into_par_iter()
                 .map(|idx| {
                     // Split file reader over index range
                     let mut file = &content_file;
@@ -194,6 +219,21 @@ impl Download {
         .await
         .expect("join handle")
     }
+}
+
+/// Returns true if all assets already exist in the installation
+async fn check_assets_exist(indicies: &[&payload::Index], installation: &Installation) -> bool {
+    stream::iter(indicies)
+        .map(|index| async move {
+            if let Ok(path) = asset_path(installation, &format!("{:02x}", index.digest)).await {
+                return fs::try_exists(path).await.unwrap_or_default();
+            }
+
+            false
+        })
+        .buffer_unordered(CONCURRENT_FS_TASKS)
+        .all(|exists| async move { exists })
+        .await
 }
 
 pub async fn download_path(installation: &Installation, hash: &str) -> Result<PathBuf, Error> {
